iommu/arm-smmu: Add SMMU workaround for Phytium Ps17064
There are two hacks needed for IOMMU enablement on Phytium
Ps17064 SoC. One is the MSI hack, the other is the SMMU hack.

When using this enablement, we assumes that users would set
CONIFIG_IOMMU_DEFAULT_PASSTHROUGH=y or pass 'iommu.passthrough=on'
or 'iommu.pt' as the kernel command-line parameters. Therefore,
we also force default iommu domain type to IOMMU_DOMAIN_IDENTITY
on Ps17064 to avoid unnecessary troubles.

Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>
Signed-off-by: Chen Baozi <chenbaozi@phytium.com.cn>
Change-Id: I60a83f56b791e36fab8b1cfec62a760dd70d1f81