net: macb: Allocate TX/RX rings in the DMA32 zone
The controller implementation shares a single register for the
high address bits across all TX/RX rings. If the buffers are
allocated in different 4GB ranges, DMA on the queues would fail.

This patch fixes the issue by allocating the ring buffers in DMA32
memory zone which makes the allocation before 4G memory.

Signed-off-by: Chen Baozi <chenbaozi@phytium.com.cn>
Change-Id: I521961829e14d90165f166667ad8b6f15dc12150